chore: SyncModalAction and useSyncModal
