+++ packages/frontend/src/features/sync/components/SyncModalDelete.tsx
@@ -4,7 +4,8 @@ import ErrorState from '../../../components/common/ErrorState';
 import SuboptimalState from '../../../components/common/SuboptimalState/SuboptimalState';
 import { useScheduler } from '../../../features/scheduler/hooks/useScheduler';
 import { useSchedulersDeleteMutation } from '../../../features/scheduler/hooks/useSchedulersDeleteMutation';
-import { SyncModalAction, useSyncModal } from '../providers/SyncModalProvider';
+import { useSyncModal } from '../providers/useSyncModal';
+import { SyncModalAction } from '../types';
 
 export const SyncModalDelete = () => {
     const { currentSchedulerUuid, setAction } = useSyncModal();
