+++ packages/frontend/src/features/sync/providers/useSyncModal.tsx
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { SyncModalContext } from './SyncModalProvider';
+
+export const useSyncModal = () => {
+    const context = useContext(SyncModalContext);
+    if (!context) {
+        throw new Error(
+            'useSyncWithGoogleSheets must be used within a SyncModalProvider',
+        );
+    }
+    return context;
+};
