+++ packages/frontend/src/features/sync/types/index.ts
@@ -1,0 +1,6 @@
+export enum SyncModalAction {
+    CREATE = 'create',
+    EDIT = 'edit',
+    VIEW = 'view',
+    DELETE = 'delete',
+}
